rename and restructure vase overview components
