+++ plugin-hrm-form/src/___tests__/components/case/EditCaseSummary.test.tsx
@@ -26,7 +26,7 @@ import { DefinitionVersion, DefinitionVersionId, loadDefinition } from 'hrm-form
 
 import { mockLocalFetchDefinitions } from '../../mockFetchDefinitions';
 import { mockGetDefinitionsResponse } from '../../mockGetConfig';
-import EditCaseSummary, { EditCaseSummaryProps } from '../../../components/case/caseDetails/EditCaseSummary';
+import EditCaseSummary, { EditCaseSummaryProps } from '../../../components/case/caseDetails/EditCaseOverview';
 import { getDefinitionVersions } from '../../../hrmConfig';
 import { StandaloneITask } from '../../../types/types';
 import { CaseRoute } from '../../../states/routing/types';
