+++ plugin-hrm-form/src/components/case/CaseTags.tsx
@@ -20,11 +20,11 @@ import React from 'react';
 import { View, Text } from '@react-pdf/renderer';
 import { DefinitionVersion } from 'hrm-form-definitions';
 
-import { Flex } from '../../../styles';
-import { TagsWrapper } from '../../search/styles';
-import CategoryWithTooltip from '../../common/CategoryWithTooltip';
-import { getContactTags } from '../../../utils/categories';
-import styles from './styles';
+import { Flex } from '../../styles';
+import { TagsWrapper } from '../search/styles';
+import CategoryWithTooltip from '../common/CategoryWithTooltip';
+import { getContactTags } from '../../utils/categories';
+import styles from './casePrint/styles';
 
 type OwnProps = {
   categories?: {
